requirements.txt
# pytorch==2.2.0    # installed through conda
# cython<3.0.0
# pyyaml==5.4.1
argparse==1.4.0
tqdm==4.66.5
nemo_toolkit @ git+https://github.com/AI4Bharat/NeMo@nemo-v2
transformers==4.40.0
huggingface_hub==0.23
pytorch-lightning==2.4.0
hydra-core==1.3.2
librosa==0.10.2.post1
sentencepiece==0.2.0
pandas==2.2.2
lhotse==1.27.0
editdistance==0.8.1
jiwer==3.0.4
pyannote.audio==3.3.1
webdataset==0.2.100
datasets==2.21.0
IPython==8.27.0
sonar-space==0.2.1
joblib==1.4.2
minio==7.2.8
argparse==1.4.0
inflect==7.4.0
fairseq2==0.2.1
fairseq2n==0.2.1
pydub==0.25.1
glob2==0.7
indic-nlp-library==0.92